+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
   * Simulcast send attributes RFC 8851 #780
+  * Set SDP port 0 for rejected m-lines #772
   * Handle incoming SSRC 0 BWE probe #769
   * New CryptoProvider API #760
   * Update H.264 profile and level matching logic #758